Split the two longest test files so they spread across test workers
ParallelTestRunner schedules whole files, so the wall time floor of a test
job is the longest single file. On CI that is linalg.jl (348 s on ubuntu,
920 s on macOS) followed by cholmod.jl (316 s / 900 s), while the other
workers idle at the end.

Move the product, dot, Kronecker and conversion testsets of linalg.jl into
linalg_products.jl, and the core Sparse/Dense/Factor operation and
regression testsets of cholmod.jl into cholmod_ops.jl. Each new file
carries the same prelude as its source (the CHOLMOD one keeps the outer
index/element type loop). No testset is changed. Measured locally, the
halves take 56 s + 44 s (linalg) and 18 s + 53 s (cholmod).
